feat: Add getRolesServices function to retrieve roles services
This commit adds a new function called getRolesServices to the default.js file in the services/ministries directory. This function retrieves all the roles services from the roles_services table in the database. It executes a SELECT query and returns the result rows.

The getRolesServices function is also implemented in the default.js file in the controllers/ministries directory. It is used to handle the GET request for the /services endpoint in the router.

This commit introduces a new database query and controller function to support the retrieval of roles services.